requirements.txt
atomicwrites==1.3.0
attrs==19.2.0
bcrypt==3.1.7
-e git+https://github.com/justinmoon/bedrock.git@ba39ebea0e54543487d18a192b8861824b4d8cf0#egg=Bedrock
blinker==1.4
cachetools==3.1.1
certifi==2019.9.11
cffi==1.13.0
chardet==3.0.4
Click==7.0
ecdsa==0.13.2
entrypoints==0.3
esptool==2.7
flake8==3.7.8
Flask==1.1.1
Flask-Bcrypt==0.7.1
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.2
google-api-python-client==1.7.11
google-auth==1.6.3
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
httplib2==0.14.0
idna==2.8
importlib-metadata==0.23
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
mccabe==0.6.1
more-itertools==7.2.0
oauthlib==3.1.0
packaging==19.2
pbkdf2==1.3
phonenumbers==8.10.20
Pillow==6.2.0
pluggy==0.13.0
py==1.8.0
pyaes==1.6.1
pyasn1==0.4.7
pyasn1-modules==0.2.7
pycodestyle==2.5.0
pycparser==2.19
pyflakes==2.1.1
PyJWT==1.7.1
pyparsing==2.4.2
pyserial==3.4
PySocks==1.7.1
pytest==5.2.1
python-bitcoinrpc==1.0
pytz==2019.3
pyudev==0.21.0
requests==2.22.0
requests-oauthlib==1.2.0
rsa==4.0
rshell==0.0.26
six==1.12.0
SQLAlchemy==1.3.10
svgwrite==1.3.1
Tree==0.2.4
twilio==6.31.1
uritemplate==3.0.0
urllib3==1.25.6
wcwidth==0.1.7
Werkzeug==0.16.0
WTForms==2.2.1
zipp==0.6.0
